requirements.txt
blis==0.2.4
certifi==2019.3.9
chardet==3.0.4
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
cymem==2.0.2
futures==2.1.6
idna==2.8
jsonschema==2.6.0
lazy-streams==0.4
mathparse==0.1.2
murmurhash==1.0.2
mysql-connector==2.2.9
nltk==3.4.1
numpy==1.16.3
Pint==0.9
plac==0.9.6
preshed==2.0.1
promise-keeper==0.3
pymongo==3.8.0
pystreams==0.6
python-dateutil==2.7.5
pytz==2019.1
PyYAML==5.1
repoze.lru==0.6
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.6.1
spacy==2.1.4
SQLAlchemy==1.2.19
srsly==0.0.5
thinc==7.0.4
tqdm==4.31.1
urllib3==1.24.3
wasabi==0.2.1
